Update README.md
The code above is written in a very simple way in the R programing language. This code shows you how to develop machine learning models using multiple algorithms.
The best part of this code is to visualize all the model plots in a single graph. I have tried to cover most of the models in this code. Please feel free if you have any query.